requirements.txt
twilio
python-dotenv
sqlalchemy
mysql-connector-python
fastapi
uvicorn[standard]
python-multipart
fastapi[all]
passlib[bcrypt]
python-jose
